+++ packages/core/src/components/header/header.stories.ts
@@ -1,0 +1,15 @@
+import { storiesOfHeader } from "@dso-toolkit/sources";
+import { storiesOf } from "@storybook/web-components";
+import readme from "./readme.md";
+import { headerTemplate } from "./header.template";
+
+storiesOfHeader(
+  {
+    module,
+    storiesOf,
+    readme,
+  },
+  {
+    headerTemplate,
+  }
+);